ITS#8015: replace memcpy() by memmove() for overlapped src & dst.
The "rebus" codestyle:
 - AC_MEMCPY() is used in many places of source code;
 - memcpy() and memmove() are used too;
 - but in real AC_MEMCPY is a memmove()...
 - but if memmove() is available...

This is really crazy, because memcpy() or bcopy() could be
SILENTLY used for overlapped source and destination regions.
http://lwn.net/Articles/414467/

Initially this wonder-bug was discovered in such situation
as a few cases by using Valgrind. After that I manually reviewed
all calls memcpy() and AC_MEMCPY() it the 2.4 branch of original OpenLDAP.

Unfortunately, we can not expect that all of these mistakes
corrected, but it is much more than nothing.

The next step is reviewing all AC_MEMCPY() calls again and
COMPLETELY replace by memmove() or memcpy().

Change-Id: I32a67a170788687b7197f6fbcad19838541f73d3