contrib-modules: ITS#8097 nssov update to protocol version 2.
This updates nssov for the protocol changes in nss-pam-ldapd commits
5f55781 and 6a74d8d. The protocol was changed to network byte order,
uid_t and gid_t were changed to int32_t, and the READ_TYPE and
WRITE_TYPE macros were removed. The PAM protocol was restructured to
drop the DN field and to use a common basic set of fields for all
requests.

Change-Id: I40d69fff4cabc8e8974810aa3257914f84bbe253