slapd: ITS#8108 silence rootdn check on hidden backends.
select_backend() skips hidden backends so it fails to match the
rootdn of a database that has been hidden. rootpw tries to see
if the rootdn matches the current backend but the check fails
when hidden. Ignore this check on hidden backends since one
cannot Bind to a hidden backend anyway.

Change-Id: Id256265968885eb0aef99a9ea5b20f8a67006622